refactor(roadmap): split sync.ts along responsibility seams
The `required-review` gate flagged `roadmap/sync.ts` at 441 lines against a
300-line threshold. Split along real seams rather than an arbitrary cut,
following the `triage.ts` + `triage-*.ts` sibling convention already used in
this directory:

- `sync.ts` (213) — thin entry point: commander definition, flag mapping,
  orchestration.
- `sync-report.ts` (143) — the report shape, its `--json` projection, and its
  rendering.
- `sync-verdict.ts` (73) — the pass / abstain / fail decision and the exit-code
  contract.
- `sync-deps.ts` (77) — resolving the tracker config and adapter from the
  environment.

Behaviour is identical: no interface, output, or exit-code change. Existing
tests pass unmodified, and the live probes reproduce byte-identical prose, the
same `--json` payload at both flag positions, and the same exit codes.
`RoadmapSyncReport` and `logSyncReport` are re-exported from `./sync` so it
remains the single import site for the command's public surface.

The same gate finding also carried a "critical / CWE-89 SQL injection" label
whose rationale, suggestion, and evidence were entirely about file length
(citing a commander `.option()` help string). There is no SQL in this file; that
finding-aggregation bug is filed upstream as #984. Only the file-length half was
real, and this addresses it.
